created progress bar function
